Use correct "ByteCommandParameter" for three parameters
Using "UshortCommandParameters" for advertisingType, ownAdressType and peerAdressType caused 3 "00"-Bytes extra within the parameter, causing the parameter to be ignored by the Bluetooth device. Changing this to the correct "ByteCommandParameter", the parameter holds 16 Bytes which are accepted by the Bluetooth device. Only by employing this change, the advertisment-Interval can be changed at all.